split out persist inbound and rename
