+++ src/repos/comms-message.js
@@ -2,7 +2,6 @@ import { config } from '../config/index.js'
 import db from '../data/db.js'
 
 const notificationsCollection = config.get('mongo.collections.notifications')
-const fileMetadataCollection = config.get('mongo.collections.fileMetadata')
 
 const persistCommsNotification = async (notification) => {
   try {
@@ -12,15 +11,6 @@ const persistCommsNotification = async (notification) => {
   }
 }
 
-const persistFileMetadata = async (metadata) => {
-  try {
-    await db.collection(fileMetadataCollection).insertOne(metadata)
-  } catch (error) {
-    throw new Error(`Error while persisting file metadata: ${error.message}`)
-  }
-}
-
 export {
-  persistCommsNotification,
-  persistFileMetadata
+  persistCommsNotification
 }
