+++ test/unit/messaging/inbound/data-ingest/processors/comms.test.js
@@ -16,7 +16,7 @@ jest.unstable_mockModule('../../../../../../src/logging/logger.js', () => ({
   })
 }))
 
-jest.unstable_mockModule('../../../../../../src/interfaces/persist-inbound.js', () => ({
+jest.unstable_mockModule('../../../../../../src/repos/comms-message.js', () => ({
   persistCommsNotification: mockPersistCommsNotification
 }))
